refactor: For each js one css && Apply BEM to css
