+++ src/components/Header/Header.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import SearchBar from './SearchBar';
 
+import './Header.css';
+
 class Header extends React.Component {
 
     render() {
