+++ src/components/profile/Gallery/Gallery.css
@@ -3,20 +3,3 @@
     flex-wrap: wrap;
 }
 
-.gallery__item {
-    position: relative;
-    flex: 1 0 250px;
-    margin: 1rem;
-    color: #fff;
-    cursor: pointer;
-    width: 293px;
-    height: 293px;
-
-}
-
-.gallery__item img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
