+++ src/components/profile/Gallery/GalleryItem.css
@@ -1,0 +1,16 @@
+.gallery-item {
+    position: relative;
+    flex: 1 0 250px;
+    margin: 1rem;
+    color: #fff;
+    cursor: pointer;
+    width: 293px;
+    height: 293px;
+
+}
+
+.gallery-item img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
