+++ src/components/profile/Header/Header.css
@@ -1,17 +1,3 @@
-.profile__header {
+.header__dashboard {
     display: flex;
-}
-
-.profile__photo {
-    flex-grow: 1;
-}
-
-.profile__photo img {
-    width: 160px;
-    display: block;
-    margin: 0 auto;
-}
-
-.profile__information {
-    flex-grow: 3;
 }
