+++ src/components/profile/Header/Header.js
@@ -1,16 +1,18 @@
 import React from 'react';
-import Photo from "./Photo";
-import Information from "./Information";
-import Stories from "./Stories/Stories";
+
+import Icon from './Icon/Icon';
+import Information from './Information/Information';
+import Stories from './Stories/Stories';
+
 import "./Header.css";
 
 
 class Header extends React.Component {
     render() {
         return(
-            <div>
-               <div className="profile__header">
-                   <Photo />
+            <div className="header">
+               <div className="header__dashboard">
+                   <Icon />
                    <Information />
                </div>
