+++ src/components/profile/Header/Icon/Icon.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import './Icon.css';
+
+class Icon extends React.Component {
+    render() {
+        return(
+            <div className="icon">
+                <img src="logo192.png"  alt="Icon"/>
+            </div>
+        );
+    }
+}
+
+export default Icon;
