+++ src/components/profile/Header/Information/Information.css
@@ -1,0 +1,3 @@
+.information {
+    flex-grow: 5;
+}
