+++ src/components/profile/Header/Stories/Stories.css
@@ -1,27 +1,4 @@
-.profile__stories {
+.stories {
     display: flex;
     margin: 70px 0;
-}
-
-.story {
-    width: 125px;
-    font-size: 14px;
-    flex: 0 0 auto;
-    text-align: center;
-}
-
-.story .story__border {
-    width: 82px;
-    height: 82px;
-    border: 1px solid #cdcdcd;
-    border-radius: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin: 0 auto;
-}
-
-.story img {
-    width: 75px;
-    border-radius: 100%;
 }
