+++ src/components/profile/Header/Stories/Story.js
@@ -1,4 +1,6 @@
-import React from "react";
+import React from 'react';
+
+import './Story.css';
 
 class Story extends React.Component {
 
@@ -6,7 +8,7 @@ class Story extends React.Component {
         return(
             <div className="story">
                 <div className="story__border">
-                    <img src={this.props.src} alt="Photo"/>
+                    <img src={this.props.src} alt="Icon"/>
                 </div>
                 <p><b>{this.props.title}</b></p>
             </div>
